Migration working

+++ elyko/database/migrations/2016_03_21_231922_create_uvs_table.php
@@ -10,7 +10,7 @@ class CreateUvsTable extends Migration {
 		Schema::create('uvs', function(Blueprint $table) {
 			$table->integer('id')->unsigned();
 			$table->string('name');
-			$table->integer('credits');
+			$table->float('credits');
 			$table->string('semester');
 			$table->primary('id');
 		});
